// frontend/src/api/models/notificationInboxItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { NotificationInboxItemType } from './notificationInboxItemType'

export interface NotificationInboxItem {
  id: string
  type: NotificationInboxItemType
  title: string
  body: string
  url: string
  createdAt: string
  /** @nullable */
  readAt: string | null
}